Replace the current list implementation in the library
The current list implementation used in the library is borrowed
from the Linux kernel sources and is released under GPL, leading to
a contamination inside the library.
To avoid this, this patch drops the current list implementation and
provides a new one written from scratch which exposes a quite similar
interface (yet slightly different to better fit our needs).
For this occasion, the code of this implementation moves from common/ to
libopen-mx/.
(this commit corresponds to a cherry pick from the both following ones:
 - 0c3a23c3c873b7aadd3c55e12a1b8c4a62b841f7
 - 7c80cfd107d3906b89737b8284f52099ae3b6248)