Modified again the --with-linux-release behavior
The commit 4a9ee4fb75837f805f68317e22b10d1574300af1 made the
--with-linux-release option mandatory when specifying --with-linux
and/or --with-linux-build.
This was done to work around some buggy kernels with 'make
kernelrelease'.

But a lot of kernels deal correctly with 'make kernelrelease' and it is
convenient to just specify the Linux build tree location.

So, this commit reverts some part of the above commit by making
--with-linux-release optional when specifying the build tree.

The Linux configuration policy is now the following :

1) If nothing is given, all is good because `uname -r` is used
everywhere.

2) If --with-linux-release is given, it is used as final kernel release.
We just check if this value matches the 'make kernelrelease' one and
issue a minor warning else.

3) --with-linux-release is not given but the build tree is (through
--with-linux and/or --with-linux-release). We have no choice of taking
the 'make kernelrelease' value but we issue a warning at the end of the
configure to inform the user with these famous buggy kernels.
The workaround in this case is to force --with-linux-release