Fix build when the kernel has DMA_ENGINE without NET_DMA.
CONFIG_DMA_ENGINE is enough to get channels and so on
on 2.6.29+ kernels.
But CONFIG_NET_DMA is still required since we use skb->dma_cookie.

Never noticed earlier because NET_DMA is basically always enabled
on x86 platform since the IOATDMA driver is enabled.

Thanks to Sébastien Varrette for reporting the problem.